requirements.txt
argcomplete==1.10.0
certifi==2019.6.16
chardet==3.0.4
ConfigArgParse==0.14.0
DateTime==4.3
dnspython==1.16.0
idna==2.8
mysql-connector==2.2.9
mysql-connector-python==8.0.16
oneagent-sdk==1.2.1
pings==0.0.1
protobuf==3.8.0
pymongo==3.8.0
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
zope.interface==4.6.0
